+++ example/lib/pages/home.dart
@@ -1,15 +1,20 @@
 import 'package:dart_web/dart_web.dart';
 
-import 'button.dart';
-import 'counter.dart';
+import '../components/button.dart';
+import '../components/counter.dart';
 
 class Home extends StatelessComponent {
   @override
   Iterable<Component> build(BuildContext context) sync* {
+    yield DomComponent(
+      tag: 'span',
+      child: Text('ABOUT'),
+    );
+
     yield Counter();
 
     yield Button(
-      label: 'Go To About',
+      label: 'About',
       onPressed: () {
         Router.of(context).push('/about', eager: true);
       },