+++ lib/src/framework/app_binding.dart
@@ -1,5 +1,6 @@
 part of framework;
 
+/// Main app binding, controls the root component and global state
 abstract class AppBinding {
   Uri get currentUri;
 
@@ -9,15 +10,6 @@ abstract class AppBinding {
   AppBinding() {
     _instance = this;
     _stateData = loadStateData();
-  }
-
-  bool get isFirstBuild => _buildQueue.isNotEmpty;
-
-  final List<Future> _buildQueue = [];
-  Future<void> get firstBuild async {
-    while (_buildQueue.isNotEmpty) {
-      await _buildQueue.first;
-    }
   }
 
   void attachRootComponent(Component app, {required String to}) {
@@ -36,6 +28,7 @@ abstract class AppBinding {
   @protected
   void attachRootElement(BuildScheduler element, {required String to}) {}
 
+  /// Global state data returned from the server. See [PreloadStateMixin]
   late final Map<String, dynamic> _stateData;
 
   @protected
@@ -53,6 +46,8 @@ abstract class AppBinding {
 
   bool isLoadingState = false;
 
+  /// Notifies elements about new global state synced from the server
+  /// This is called when a [LazyRoute] is eagerly loaded.
   void notifyState(Map<String, dynamic> data) {
     isLoadingState = false;
     for (var id in data.keys) {
@@ -69,6 +64,15 @@ abstract class AppBinding {
     }
   }
 
+  bool get isFirstBuild => _buildQueue.isNotEmpty;
+
+  final List<Future> _buildQueue = [];
+  Future<void> get firstBuild async {
+    while (_buildQueue.isNotEmpty) {
+      await _buildQueue.first;
+    }
+  }
+
   void performRebuildOn(Element? child) {
     var built = performRebuild(child);
     if (built is Future) {
@@ -78,6 +82,9 @@ abstract class AppBinding {
     }
   }
 
+  /// The first build on server and browser is allowed to have asynchronous operations (i.e. preloading data)
+  /// However we want the component and element apis to stay synchronous, so subclasses
+  /// can override this method to simulate an async rebuild for a child
   @protected
   @mustCallSuper
   FutureOr<void> performRebuild(Element? child) {
@@ -94,6 +101,33 @@ abstract class AppBinding {
     if (_globalKeyRegistry[key] == element) {
       _globalKeyRegistry.remove(key);
     }
+  }
+}
+
+/// In difference to Flutter, we have multiple build schedulers instead of one global build owner
+/// Particularly each dom element is a build scheduler and manages its subtree of components
+mixin BuildScheduler on Element {
+  DomView? view;
+
+  Future? _rebuilding;
+
+  // TODO: find out how activation / deactivation of a build scheduler affects its child elements, this probably has some bugs still
+  final _InactiveElements _inactiveElements = _InactiveElements();
+
+  void scheduleBuildFor(Element element) {
+    scheduleRebuild();
+  }
+
+  Future<void> scheduleRebuild() {
+    return _rebuilding ??= Future.microtask(() async {
+      try {
+        rebuild();
+        view?.update();
+        _inactiveElements._unmountAll();
+      } finally {
+        _rebuilding = null;
+      }
+    });
   }
 }
 
@@ -114,28 +148,4 @@ class _RootElement extends SingleChildElement with BuildScheduler {
 
   @override
   Component build() => component.child;
-}
-
-mixin BuildScheduler on Element {
-  DomView? view;
-
-  Future? _rebuilding;
-
-  final _InactiveElements _inactiveElements = _InactiveElements();
-
-  void scheduleBuildFor(Element element) {
-    scheduleRebuild();
-  }
-
-  Future<void> scheduleRebuild() {
-    return _rebuilding ??= Future.microtask(() async {
-      try {
-        rebuild();
-        view?.update();
-        _inactiveElements._unmountAll();
-      } finally {
-        _rebuilding = null;
-      }
-    });
-  }
 }