+++ lib/src/framework/build_context.dart
@@ -1,14 +1,12 @@
 part of framework;
 
+/// As known from Flutter
 abstract class BuildContext {
-  InheritedComponent dependOnInheritedElement(InheritedElement ancestor,
-      {Object? aspect});
+  InheritedComponent dependOnInheritedElement(InheritedElement ancestor, {Object? aspect});
 
-  T? dependOnInheritedComponentOfExactType<T extends InheritedComponent>(
-      {Object? aspect});
+  T? dependOnInheritedComponentOfExactType<T extends InheritedComponent>({Object? aspect});
 
-  InheritedElement? getElementForInheritedComponentOfExactType<
-      T extends InheritedComponent>();
+  InheritedElement? getElementForInheritedComponentOfExactType<T extends InheritedComponent>();
 
   T? getInheritedComponentOfExactType<T extends InheritedComponent>();