+++ lib/src/framework/dom_component.dart
@@ -2,6 +2,7 @@ part of framework;
 
 typedef EventCallback = void Function();
 
+/// Represents a dom element
 class DomComponent extends Component {
   DomComponent({
     Key? key,
@@ -66,6 +67,8 @@ class DomElement extends SingleChildElement with BuildScheduler {
   }
 }
 
+/// Represents a plain text node
+/// This has no other properties. I.e. styling is done through the parent element(s) and their styles.
 class Text extends Component {
   Text(this.text, {Key? key}) : super(key: key);