+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0
+
+- Completed component & element structure
+- Added router with lazy loading routes
+
 ## 0.0.1
 
 - Initial development release